+++ TombEngine/Objects/TR3/Entity/tr3_mp_stick.cpp
@@ -155,7 +155,7 @@ namespace TEN::Entities::Creatures::TR3
 
 			angle = CreatureTurn(item, creature->MaxTurn);
 
-			auto* enemy = creature->Enemy;
+			auto* enemy = creature->Enemy.Get();
 			creature->Enemy = LaraItem;
 
 			if (item->HitStatus || ((laraAI.distance < pow(BLOCK(1), 2) ||
@@ -343,7 +343,7 @@ namespace TEN::Entities::Creatures::TR3
 					extraTorsoRot.y = AI.angle;
 				}
 
-				if (creature->Enemy->IsLara())
+				if (creature->Enemy.IsLara())
 				{
 					if (!creature->Flags && item->TouchBits.Test(MPStickPunchAttackJoints))
 					{
@@ -378,7 +378,7 @@ namespace TEN::Entities::Creatures::TR3
 					extraTorsoRot.y = AI.angle;
 				}
 
-				if (creature->Enemy->IsLara())
+				if (creature->Enemy.IsLara())
 				{
 					if (!creature->Flags && item->TouchBits.Test(MPStickPunchAttackJoints))
 					{
@@ -416,7 +416,7 @@ namespace TEN::Entities::Creatures::TR3
 					extraTorsoRot.y = AI.angle;
 				}
 
-				if (creature->Enemy->IsLara())
+				if (creature->Enemy.IsLara())
 				{
 					if (creature->Flags != 2 && item->TouchBits.Test(MPStickPunchAttackJoints))
 					{
@@ -448,7 +448,7 @@ namespace TEN::Entities::Creatures::TR3
 				if (AI.ahead)
 					extraTorsoRot.y = AI.angle;
 
-				if (creature->Enemy->IsLara())
+				if (creature->Enemy.IsLara())
 				{
 					if (creature->Flags != 1 && item->TouchBits.Test(MPStickKickAttackJoints) &&
 						item->Animation.FrameNumber > 8)