+++ TombEngine/Objects/TR4/Entity/tr4_dog.cpp
@@ -119,7 +119,7 @@ namespace TEN::Entities::TR4
 			CreatureAIInfo(item, &AI);
 
 			int distance;
-			if (creature->Enemy->IsLara())
+			if (creature->Enemy.IsLara())
 			{
 				distance = AI.distance;
 			}